Select frame where occurrence appears for the first time
